Revert "FC Diff view (FC section only)  (#2569)"
This reverts commit 8572014b4eb60763ea456160abaa3c890d803dc6.
