+++ src/components/v1/shared/FundingCycle/QueuedFundingCycle.tsx
@@ -37,9 +37,7 @@ export default function QueuedFundingCycle() {
           </CardSection>
         </div>
       ) : (
-        <div className="text-gray-500 dark:text-slate-100">
-          No upcoming funding cycle
-        </div>
+        <div className="text-gray-500 dark:text-slate-100">No upcoming funding cycle</div>
       )}
     </div>
   )
