+++ src/components/v2v3/V2V3Project/V2V3FundingCycleSection/FundingCycleDetails/FundingCycleListItem.tsx
@@ -1,42 +1,10 @@
 import { Tooltip } from 'antd'
-import { twJoin } from 'tailwind-merge'
 import { classNames } from 'utils/classNames'
 
-// whether this value the old value or a new (updated) value
-type DiffStatus = 'new' | 'old'
-
-function ListItemValue({
-  value,
-  diffStatus,
-}: {
-  value: string | JSX.Element
-  diffStatus?: DiffStatus
-}) {
-  const highlight =
-    diffStatus === 'old'
-      ? 'bg-error-100 dark:bg-error-900'
-      : diffStatus === 'new'
-      ? 'bg-success-100 dark:bg-success-900'
-      : undefined
-
+function ListItemValue({ value }: { value: string | JSX.Element }) {
   return (
-    <div
-      className={twJoin(
-        'text-secondary ml-2 flex whitespace-nowrap',
-        highlight,
-        diffStatus ? 'pr-1' : undefined,
-      )}
-    >
-      {diffStatus ? (
-        <span className="mr-2">
-          {diffStatus === 'new' ? (
-            <span className="text-success-500 dark:text-success-200">+</span>
-          ) : diffStatus === 'old' ? (
-            <span className="text-error-500 dark:text-error-200">–</span>
-          ) : null}
-        </span>
-      ) : null}
-      <div className={diffStatus ? 'font-medium' : undefined}>{value}</div>
+    <div className="ml-1 whitespace-nowrap text-grey-500 dark:text-grey-300">
+      {value}
     </div>
   )
 }
@@ -46,24 +14,13 @@ export function FundingCycleListItem({
   name,
   helperText,
   value,
-  oldValue,
   subItem,
 }: {
   name: string
   value: string | JSX.Element
-  oldValue?: string | JSX.Element
   helperText?: string | JSX.Element
   subItem?: boolean
 }) {
-  const hasDiff = oldValue && value !== oldValue
-
-  const _value = (
-    <>
-      {hasDiff ? <ListItemValue value={oldValue} diffStatus={'old'} /> : null}
-      <ListItemValue value={value} diffStatus={hasDiff ? 'new' : undefined} />
-    </>
-  )
-
   if (helperText) {
     return (
       <div
@@ -85,7 +42,8 @@ export function FundingCycleListItem({
             :
           </div>{' '}
         </Tooltip>
-        {_value}
+
+        <ListItemValue value={value} />
       </div>
     )
   }
@@ -97,7 +55,7 @@ export function FundingCycleListItem({
         subItem ? 'ml-5 text-xs' : 'text-sm',
       )}
     >
-      <div className="font-medium">{name}:</div> {_value}
+      <div className="font-medium">{name}:</div> <ListItemValue value={value} />
     </div>
   )
 }
