--- src/components/v2v3/V2V3Project/V2V3FundingCycleSection/FundingCycleDetails/RulesListItems/AllowSetTerminalsValue.tsx
@@ -1,9 +1,0 @@
-import { Trans } from '@lingui/macro'
-
-export function AllowSetTerminalsValue({
-  allowSetTerminals,
-}: {
-  allowSetTerminals: boolean
-}) {
-  return allowSetTerminals ? <Trans>Allowed</Trans> : <Trans>Disabled</Trans>
-}
