--- src/components/v2v3/V2V3Project/V2V3FundingCycleSection/FundingCycleDetails/RulesListItems/HoldFeesValue.tsx
@@ -1,5 +1,0 @@
-import { Trans } from '@lingui/macro'
-
-export function HoldFeesValue({ holdFees }: { holdFees: boolean }) {
-  return holdFees ? <Trans>Enabled</Trans> : <Trans>Disabled</Trans>
-}
