--- src/components/v2v3/V2V3Project/V2V3FundingCycleSection/FundingCycleDetails/TokenListItems/PayerOrReservedTokensValue.tsx
@@ -1,18 +1,0 @@
-import { Trans } from '@lingui/macro'
-import { formattedNum } from 'utils/format/formatNumber'
-
-export function PayerOrReservedTokensValue({
-  value,
-  tokenSymbol,
-}: {
-  value: string
-  tokenSymbol: string
-}) {
-  return (
-    <span>
-      <Trans>
-        {formattedNum(value)} {tokenSymbol}/ETH
-      </Trans>
-    </span>
-  )
-}
