--- src/components/v2v3/V2V3Project/V2V3FundingCycleSection/FundingCycleDetails/TokenListItems/ReservedRateValue.tsx
@@ -1,22 +1,0 @@
-import { BigNumber } from '@ethersproject/bignumber'
-import FundingCycleDetailWarning from 'components/Project/FundingCycleDetailWarning'
-import { FUNDING_CYCLE_WARNING_TEXT } from 'constants/fundingWarningText'
-import { formatReservedRate } from 'utils/v2v3/math'
-
-export function ReservedRateValue({
-  value,
-  showWarning,
-}: {
-  value: BigNumber
-  showWarning?: boolean
-}) {
-  const riskWarningText = FUNDING_CYCLE_WARNING_TEXT()
-  return (
-    <FundingCycleDetailWarning
-      showWarning={showWarning}
-      tooltipTitle={riskWarningText.metadataReservedRate}
-    >
-      {formatReservedRate(value)}%
-    </FundingCycleDetailWarning>
-  )
-}
