+++ src/utils/v2v3/math.ts
@@ -33,8 +33,6 @@ export const MAX_MINT_RATE = Math.floor(MaxUint88 / 10 ** 18)
 
 const MAX_FEE = ONE_BILLION
 
-type IssueeType = 'reserved' | 'payer'
-
 /**
  * Express a given discount rate (parts-per-billion) as a percentage.
  * @param discountRate - discount rate as parts-per-billion.
@@ -216,7 +214,7 @@ export const weightAmountPermyriad: WeightFunction = (
   weight: BigNumber | undefined,
   reservedRatePermyriad: number | undefined,
   amountWad: BigNumber | undefined,
-  outputType: IssueeType,
+  outputType: 'payer' | 'reserved',
 ): string => {
   if (!weight || !amountWad) return '0'
 
@@ -264,10 +262,9 @@ export function sumHeldFees(fees: JBFee[]) {
   }, 0)
 }
 
-export function computeIssuanceRate(
+export function computePaymentIssuanceRate(
   fundingCycle: V2V3FundingCycleData,
   fundingCycleMetadata: V2V3FundingCycleMetadata,
-  issuee: IssueeType,
 ) {
   return formattedNum(
     formatIssuanceRate(
@@ -275,7 +272,7 @@ export function computeIssuanceRate(
         fundingCycle?.weight,
         fundingCycleMetadata?.reservedRate.toNumber(),
         parseEther('1'),
-        issuee,
+        'payer',
       ) ?? '',
     ),
   )
